Add Identity.

+++ backend/src/MovieStore.Domain/Entities/MovieActor.cs
@@ -3,7 +3,7 @@ namespace MovieStore.Domain.Entities;
 public class MovieActor
 {
     public int MovieId { get; set; }
-    public long ActorId { get; set; }
+    public int ActorId { get; set; }
     public string? CharacterName { get; set; }
     
     public Movie Movie { get; set; }
